chore(domain): add Stats and AddressUtxos to domain model, remove from watch-only
Move common wallet data structures `Stats` and `AddressUtxos` from
`floresta-watch-only` into the `floresta-domain` crate.
